+++ pkg/backend/fanart/bilibili/bilibili.go
@@ -20,7 +20,7 @@ func CheckNew() {
 		wg := new(sync.WaitGroup)
 		for _, Member := range database.GetMembers(Group.ID) {
 			wg.Add(1)
-			go func(Group database.GroupName, Member database.Name, wg *sync.WaitGroup) {
+			go func(Group database.Group, Member database.Member, wg *sync.WaitGroup) {
 				defer wg.Done()
 				if Member.BiliBiliHashtags != "" {
 					log.WithFields(log.Fields{