+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 v0.4.10
-- fix fanart #51
+- Fix fanart #51
+- Change some log format
+- Add http proxy in .toml config
+- Change struct name
+
 
 v0.4.9
 - Change twitter avatar scraping